Add .htaccess to specify auto-include PHP template
I might change this later to be contained within the Apache
installation's main configuration file rather than a per-directory
.htaccess file. There are some potential performance issues involved
with utilizing distributed config files and because, in some cases, I
may include .htaccess files that contain password-protected sections on
the site, or the like, and it would be strange to exclude those from git
commits while including other .htaccess files. (Sensitive .htaccess
files can be gitignored.)

The benefit about having per-directory .htaccess files is that the
contents of the site can stay separated from the Apache configuration;
if the site moves to a new server, the contents can be copied into the
new server's serving directory and the distributed configuration files
(or at least, the ones that are included in the git repo) will
immediately allow pages to be served as expected. I value this.

I feel more comfortable with and accustomed to including .htaccess files
for now, but I may change this later. If I do so, I suppose I could
document how this works in the wiki, so that if I ever need to refer to
it, there will be a record of the configuration directives affecting the
site's content and the way it's being served (so, e.g., copying the
website to a new server could be done straightforwardly, without me
having to "search" every page in the copied git repo for configuration
directives that need to be present, per-directory, but weren't commit to
github).